rename TabsSection to DifferencesSection create IntroSection create new TabsSection switcher
