+++ src/components/IntroSection.jsx
@@ -1,0 +1,10 @@
+export default function IntroSection() {
+    return (
+        <section>
+            <h1 className='centered'>Resut university</h1>
+            <h3 className='centered' style={{color:'#666'}}>
+                Пока другие предлагают 50 страниц текста и вебинары по 2 часа
+            </h3>
+        </section>
+    )
+}
