+++ src/index.css
@@ -43,3 +43,7 @@ p {
   width: 100%;
   margin-bottom: 1rem;
 }
+
+.centered{
+  text-align: center;
+}
